missing initial value h5 bug fixed
If a SPEAD item was received without initial value and
data was to be output to an h5 file, an error would be
generated.

A check is now in place to delay output to the h5 file
until a value is received. Relates in this case to
timestamp which has no value until the first SPEAD
packet arrives.