Dynamic timeout waiting for SKARAB restart.
The default timeout is currently 5 seconds in most config files. For DHCP (wait after SKARAB reboot), wait time is now calculated as  n_hosts*timeout. So, that's 40s for a 4A system and 320s (5min) for a 64A system. It may be a bit short for 4A on site, but we'll try with this first.