Fixed SyntaxError message when config file doesn't exist.
Previously, IOError was being caught when the file tries to open and fails, then
falling through to the _read_config_server function which obviously generates
a syntax error because it's a file path, not an instruction from a server.

So I rearranged the exceptions to catch this properly and report it like an
honest man.