split switch wait time into two.
We need to wait a while after resetting the fengines before trying to do a VACC sync.

So now we're resetting the fengines midway through the switch wait period.
Even if the data isn't flowing properly yet, this should be ok because the Fengines' hardware auto_rst scheme should catch it.