Add HMC debugging info to instrument monitoring loop
JIRA: CBFTASKS-784

Added additional functionality to the instrument monitoring loop. When
an HMC related error is detected a register dump of the HMC status is
grabbed and logged. Another dump is logged 2 seconds later.

The physical mezzanine site, in which each HMC is installed, is also
included in the dump.