Make descriptor string more sane.
"assert" statements all over the place are a bad idea. They don't
actually do anything in the real world, so it's a bad idea to rely on
them to sanitise inputs.

This commit removes the "assert" statement checking if the FxCorrelator
object's "descriptor" field is passed a string. In the parent Instruemnt
class, the self.descriptor field is populated with the given descriptor
string, surrounded by a try / except for just in case some idiot puts in
a non-string-thing there.

This begins a general attempt to make rampant use of these kinds of
things better, more traceable, with less duplication of effort.