Reorganised Instrument class and FxCorrelator class.
Moved logging functionality into parent Instrument class, as it makes
sense that you'd want to log whichever kind of instrument you have, not
just FxCorrelators.

Also altered the order of the arguments of Instrument's __init__()
function, to be consistent with FxCorrelator. Not that it affected
functioanlity, but some consistency is probably better.

Also modified a few of the exceptions that get raised to make them more
appropriate for the application.