Adjust response timeouts for processing nodes
A handful of changes that adjusts the response timeout when
communicating with processing nodes. The global response timeout for all
commands may now be adjusted across all processing nodes associated with
the instrument. At instrument initialisation, the global response
timeout defaults to min_load_time/2 (min_load_time is set in the config
files).

For the sensor servlet, the global response timeout is set via a new
parameter in the config files: "sensor_response_timeout". If this value
is not found in the config files, the response timeout defaults to
100 ms.

CBFTASKS-812