threaded feng write now operation
All fengines were written to in parallel, regardless of how many fengines were on a given host.
There was potential for multiple threads to try'n access the same host.
Hosts are now written in parallel, while engines on that host are addressed sequentially.

Also renamed eq_set and eq_get to set_eq and get_eq, in line with all other functions.