feat: Increase adc clip count sensor sampling rate
Science Data Processing requires a higher update of the ADC clip count
sensor so that it is at least one time per accumulation.

It is now possible to sample all clip counts from all f-engines in parallel
using fops.get_adc_clip_counts()
This new function is used to read all clip counts at the end of sensors on
every SKARAB that contains f-engines. This increases the update of the sensor
by N (where N is the number of SKARABs containing f-engines). This scales with
the number of antennas and should allow for a reasonably consistent update rate.