fix ADC sensor speedup
ADC sensors are now scheduled to be updated every second (previously
around every 10s). This works on dbelab06 for a 4 antenna correlator
although occasionally an update is skipped (once every 2s).

The automatic update rate of sensors is bypassed for these, although
kept for other sensors.

The read of these sensors is done in parallel in an attempt to
reduce overhead and impact on other scheduled tasks.

Next step is to merge and test on cmc2